twitch_bot/requirements.txt
twitchio
python-dotenv
aiofiles
